+++ R/other_utils.R
@@ -528,3 +528,78 @@ ody_apply_on_pattern <-  function(
     )
 
 }
+
+#' Label Data Frame
+#'
+#' This function labels a data frame according to a provided dictionary. It sets variable labels and value labels for specified variables within the data frame.
+#'
+#' @param raw_data A data frame containing the raw data to be labeled.
+#' @param dictionary A dictionary data frame with the following columns:
+#' - `variable`: The name of the variable to be labeled.
+#' - `variable_label` (optional): The label to assign to the variable.
+#' - `value`: The value to be labeled.
+#' - `value_label`: The label to assign to the value.
+#' Note that columns `variable` and `variable_label` will contain as many repeated values as the number of values in the `value` column.
+#'
+#' @return A data frame with labeled variables and corresponding values according to the provided dictionary.
+#' @export
+ody_label_df <- function(raw_data, dictionary) {
+
+  # Nested dictionary with vector dictionary of each variable
+  nested_dic <- dictionary |>
+    dplyr::select(.data$variable, .data$value_label, .data$value) |>
+    tidyr::nest(dic = c("value_label", "value")) |>
+    dplyr::mutate(
+      vec_dic = purrr::map_chr(
+        .data$dic,
+        ~stringr::str_c(
+          "c(",
+          stringr::str_c("`", .$value_label, "` = '", .$value, "'") |>
+            stringr::str_c(collapse = ", "),
+          ")"
+        )
+      )
+    ) |>
+    dplyr::filter(!is.na(.data$vec_dic))
+
+  # Value labels of each variable in
+  val_labels_text <- stringr::str_c("`", nested_dic$variable, "` = ", nested_dic$vec_dic) |>
+    na.omit() |>
+    stringr::str_c(collapse = ", ")
+
+  # Las variables a etiquetar han de ser caracter
+  raw_data_chr <- raw_data |>
+    dplyr::mutate(
+      dplyr::across(tidyselect::all_of(nested_dic$variable), as.character)
+    )
+
+  if (any(names(dictionary) == "variable_label")) {
+
+    var_labels_text <- dictionary |>
+      dplyr::select(.data$variable_label, .data$variable) |>
+      na.omit() |>
+      unique() |>
+      purrr::pmap(~stringr::str_c("`", .y, "`",  " = '", .x, "'")) |>
+      stringr::str_c(collapse = ", ")
+
+    labelled_vars_data <- stringr::str_c(
+      "labelled::set_variable_labels(raw_data_chr,", var_labels_text, ")"
+    ) |>
+      str2lang() |>
+      eval()
+
+  } else {
+
+    labelled_vars_data <- raw_data_chr
+
+  }
+
+  stringr::str_c(
+    "labelled::set_value_labels(labelled_vars_data,", val_labels_text, ")"
+  ) |>
+    str2lang() |>
+    eval()
+
+}
+
+