+++ inst/redcap_templates/Rprofile_template.R
@@ -3,6 +3,9 @@ source(here::here(
   list.files(here::here(), "_dependencies.R$")
 ))
 
+# Uncomment to set gander chat model
+# options(.gander_chat = ellmer::chat_openai(model = "gpt-4o-mini"))
+
 cli::rule(
   left = "Odytools REDCap Project",
   right = utils::packageVersion("odytools"),